+++ docs/fr_FR/changelog.md
@@ -1,3 +1,4 @@
+- Refonte du panel, on choisi maintenant le nombre de camera par ligne dans la configuration du plugin, leur dimension est calculé automatiquement
 - Amélioration du widget en mobile
 
 # 10/03/2018